refactor(ocr): decompose text processing logic into specialized components
The existing OCR processing logic within `AppleOCRTextProcessor` and `AppleOCRTextMerger` has been refactored to improve modularity, separation of concerns, and maintainability. Complex functionalities have been extracted into smaller, single-responsibility classes.

- **`OCRStatisticsCalculator`**: New class to handle all statistical computations for OCR observations (e.g., line height, spacing, character counts).
- **`OCRPoetryDetector`**: New class to isolate poetry detection logic.
- **`OCRDashHandler`**: New class to manage the logic for handling hyphenated words across lines.
- **`OCRTextAnalyzer`**: New class to encapsulate text analysis functions, now consumed by `AppleOCRTextMerger`.

As a result, `AppleOCRTextProcessor` is now a streamlined coordinator, delegating tasks to these new components. `AppleOCRTextMerger` is also simplified by offloading analysis to `OCRTextAnalyzer`. This restructuring makes the codebase cleaner and easier to understand and test.
