+++ Easydict/Swift/Service/Apple/AppleOCRService/OCRContext.swift
@@ -29,7 +29,7 @@ struct OCRContext {
 
 // MARK: - OCRLineContext
 
-/// Encapsulates context for a specific line of OCR text, including formatting and analysis properties
+/// Comprehensive OCR formatting data that includes all context needed for text merging
 struct OCRLineContext {
     let current: VNRecognizedTextObservation
     let previous: VNRecognizedTextObservation
@@ -53,31 +53,13 @@ struct OCRLineContext {
 
 // MARK: - OCRConstants
 
-/// Constants used for OCR text processing and merging
+/// Constants used for Apple OCR text processing
 enum OCRConstants {
-    /// Line break character
     static let lineBreakText = "\n"
-
-    /// Paragraph break characters
     static let paragraphBreakText = "\n\n"
-
-    /// Indentation text (currently empty)
-    static let indentationText = ""
-
-    /// Ratio threshold for paragraph line height detection, default is 1.5
-    static let paragraphLineHeightRatio: CGFloat = 1.5
-
-    /// Maximum character count per line for short poetry detection, default is 12
+    static let indentationCharacterCount: Double = 2.0
+    static let paragraphLineHeightRatio: Double = 1.5
     static let shortPoetryCharacterCountOfLine = 12
-
-    /// Indentation character count, default is 1.2
-    static let indentationCharacterCount = 1.2
-
-    /// Chinese difference font threshold, default is 3
-    /// Chinese fonts seem to be more precise.
-    static let chineseDifferenceFontThreshold = 3.0
-
-    /// English difference font threshold, default is 5
-    /// Note: English uppercase-lowercase font size is not precise, so threshold should a bit large.
-    static let englishDifferenceFontThreshold = 5.0
+    static let chineseDifferenceFontThreshold: Double = 3.0
+    static let englishDifferenceFontThreshold: Double = 5.0
 }
